Fixed issues with channel creation

+++ nest/backend/src/gateways/message.gateway.ts
@@ -1,6 +1,5 @@
-import { BadRequestException, HttpStatus} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { MessageBody, SubscribeMessage, WebSocketGateway, WebSocketServer, OnGatewayConnection, OnGatewayDisconnect, OnGatewayInit, ConnectedSocket } from '@nestjs/websockets';
+import { MessageBody, SubscribeMessage, WebSocketGateway, WebSocketServer, OnGatewayConnection, OnGatewayDisconnect, OnGatewayInit, ConnectedSocket, WsException } from '@nestjs/websockets';
 import {Server, Socket} from 'socket.io'
 import PostMessageDTO from 'src/dto/post-message.dto';
 import { Repository } from 'typeorm';
@@ -15,7 +14,7 @@ import { FriendMessage } from 'src/database/entities/FriendMessage';
 import CreateChannelDTO from 'src/channel/dto/create-channel.dto';
 import { ChannelMessage } from 'src/database/entities/ChannelMessage';
 import { JwtService } from '@nestjs/jwt';
-import { Logger } from '@nestjs/common';
+import { HttpStatus, Logger } from '@nestjs/common';
 import { UserService } from 'src/users/users.service';
 
 @WebSocketGateway({ namespace: 'chat' })
@@ -36,21 +35,19 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
     private sockets = new Map<Socket, number>();
 
     private verifyId(client: Socket, id: number) {
-        if (this.sockets.get(client) != id) {
-            throw new BadRequestException("Nice try");
-        }
+        if (this.sockets.get(client) != id)
+            throw new WsException("Nice try");
     }
 
     @SubscribeMessage('newmsg')
     async handleMessage(@MessageBody() dto: PostMessageDTO,
                         @ConnectedSocket() client: Socket) {
-        //this.verifyId(client, dto.senderId);
+        this.verifyId(client, dto.senderId);
         const chan: Channel = await this.channelRepository.findOneBy({id: dto.channelId})
-        if (!chan)
-            throw new BadRequestException("Channel Doesn't exist")
+        if (!chan) throw new WsException("Channel Doesn't exist")
         const user = chan.users.find(user => user.id === dto.senderId);
         if (!user || chan.isMuted(dto.senderId))
-            throw new BadRequestException("User doesn't exist, is not on this channel or is muted");
+            throw new WsException("User doesn't exist, is not on this channel or is muted");
         this.server.to(chan.id.toString()).emit(user.username + "@" + chan.id + ":" + dto.message);
         const message = this.messageRepository.create({
             content: dto.message,
@@ -65,15 +62,11 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
                         @ConnectedSocket() client: Socket) {
         this.verifyId(client, dto.uid);
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
-        if (!chan)
-            throw new BadRequestException("Channel Doesn't exist")
+        if (!chan) throw new WsException("Channel Doesn't exist")
         const user = await this.userRepository.findOneBy({id: dto.uid});
-        if (!user)
-            throw new BadRequestException("User doesn't exist");
-        if(chan.isBanned(dto.uid))
-            throw new BadRequestException("User is banned from this channel");
-        if (chan.password != null)
-            throw new BadRequestException("This channel requires a password");
+        if (!user) throw new WsException("User doesn't exist");
+        if(chan.isBanned(dto.uid)) throw new WsException("User is banned from this channel");
+        if (chan.password != null) throw new WsException("This channel requires a password");
         client.join(chan.id.toString());
         chan.users.push(user);
         this.channelRepository.save(chan);
@@ -93,15 +86,11 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
                         @ConnectedSocket() client: Socket) {
         this.verifyId(client, dto.uid);
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
-        if (!chan)
-            throw new BadRequestException("Channel Doesn't exist")
+        if (!chan) throw new WsException("Channel Doesn't exist")
         const user = await this.userRepository.findOneBy({id: dto.uid});
-        if (!user)
-            throw new BadRequestException("User doesn't exist");
-        if(chan.isBanned(dto.uid))
-            throw new BadRequestException("User is banned from this channel");
-        if (!chan.verifyPassword(dto.password))
-            throw new BadRequestException("Wrong Password");
+        if (!user) throw new WsException("User doesn't exist");
+        if(chan.isBanned(dto.uid)) throw new WsException("User is banned from this channel");
+        if (!chan.verifyPassword(dto.password)) throw new WsException("Wrong Password");
         client.join(chan.id.toString());
         chan.users.push(user);
         this.channelRepository.save(chan);
@@ -112,13 +101,10 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
     async banUser(  @MessageBody() dto: BanUserDTO,
                     @ConnectedSocket() client: Socket) {
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
-        if (!chan)
-            throw new BadRequestException("Channel Doesn't exist");
-        if (chan.admin.id != this.sockets.get(client))
-            throw new BadRequestException("Nice try");
+        if (!chan) throw new WsException("Channel Doesn't exist");
+        if (chan.admin.id != this.sockets.get(client)) throw new WsException("Nice try");
         const user = chan.users.find(user => user.id === dto.uid);
-        if (!user)
-            throw new BadRequestException("User doesn't exist or is not on this channel");
+        if (!user) throw new WsException("User doesn't exist or is not on this channel");
         if (dto.isBan) {
             const bndclient = this.clients.get(dto.uid);
             bndclient.emit("/" + chan.id + ":You were banned from this channel");
@@ -140,7 +126,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
     async getMessagePage(   @MessageBody() dto: GetMessageDTO,
                             @ConnectedSocket() client: Socket) {
         if (!client.in(dto.channelId.toString()))
-            throw new BadRequestException("Nice try");
+            throw new WsException("Nice try");
         const messages = await this.messageRepository.find({
             select: {
                 id:         true,
@@ -163,17 +149,14 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
                     @ConnectedSocket() client: Socket) {
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
         const user = await this.userRepository.findOneBy({id : dto.uid});
-        if (!chan || !user)
-            throw new BadRequestException("Channel or user doesn't exist");
-        if (chan.admin.id != this.sockets.get(client))
-            throw new BadRequestException("Nice try");
+        if (!chan || !user) throw new WsException("Channel or user doesn't exist");
+        if (chan.admin.id != this.sockets.get(client)) throw new WsException("Nice try");
         const ban = await this.bansAndMutesRepository.findOneBy({
             channel: {id: dto.chanId},
             user: {id: dto.uid}
         });
         client.emit(user.username + " was unbanned");
-        if (!ban)
-            throw new BadRequestException("User or channel doesn't exist or user is nor banned or muted from this channel");
+        if (!ban) throw new WsException("User or channel doesn't exist or user is nor banned or muted from this channel");
         this.bansAndMutesRepository.delete(ban.id);
     }
 
@@ -181,10 +164,8 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
     async deleteChannel(@MessageBody() id: number,
                         @ConnectedSocket() client: Socket) {
         const chan = await this.channelRepository.findOneBy({id})
-        if (!chan)
-            throw new BadRequestException("Channel doesn't exist");
-        if (chan.admin.id != this.sockets.get(client))
-            throw new BadRequestException("Nice try");
+        if (!chan) throw new WsException("Channel doesn't exist");
+        if (chan.admin.id != this.sockets.get(client)) throw new WsException("Nice try");
         this.server.to(chan.id.toString()).emit("/" + chan.id + ":Channel was deleted");
         this.server.socketsLeave(chan.id.toString());
         this.channelRepository.delete(id);
@@ -198,7 +179,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         let channel = this.channelRepository.create();
         const user = await this.userRepository.findOneBy({id: dto.adminId});
         if (!user)
-            throw new BadRequestException("User doesn't exist");
+            throw new WsException("User doesn't exist");
         channel.name = dto.name;
         channel.admin = user;
         channel.users = [user];
@@ -217,7 +198,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         client.emit("/" + chan.id + ":You left this channel");
         client.leave(chan.id.toString());
         if (!chan || !chan.removeUser(dto.uid))
-            throw new BadRequestException("Channel doesn't exist or user doesn't exist or is not on this channel");
+            throw new WsException("Channel doesn't exist or user doesn't exist or is not on this channel");
         this.channelRepository.save(chan);
     }
 
@@ -226,10 +207,9 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
                         @ConnectedSocket() client: Socket) {
         const message = await this.messageRepository.findOneBy({id});
         const clientid = this.sockets.get(client);
-        if (!message)
-            throw new BadRequestException("Message doesn't exist");
+        if (!message) throw new WsException("Message doesn't exist");
         if (message.sender.id != clientid && message.channel.admin.id != clientid)
-            throw new BadRequestException("Nice try");
+            throw new WsException("Nice try");
         this.server.to(message.channel.id.toString()).emit("!" + id);
         this.messageRepository.delete(id)
         return HttpStatus.NO_CONTENT;
@@ -245,7 +225,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         if (!user1 || !user2)
             return HttpStatus.NOT_FOUND;
         if (user1.blocked.includes(user2) || user2.blocked.includes(user1))
-            throw new BadRequestException("Users blocked each other");
+            throw new WsException("Users blocked each other");
         const socket = this.clients.get(user2.id);
         client.emit(user1.username + ":" + dto.message)
         if (socket) socket.emit(user1.username + ":" + dto.message)
@@ -264,7 +244,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         const message = await this.usrmessageRepository.findOneBy({id});
         const recvsocket = this.clients.get(message.receiver.id);
         if (!message || message.sender.id != this.sockets.get(client))
-            throw new BadRequestException("message dosn't exist or does not belong to the person who deleted it");
+            throw new WsException("message dosn't exist or does not belong to the person who deleted it");
         client.emit("!" + id);
         if (recvsocket) recvsocket.emit("!" + id);
         this.messageRepository.delete(id);
@@ -278,11 +258,11 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         const user1 = await this.userRepository.findOneBy({id: ids[0]});
         const user2 = await this.userRepository.findOneBy({id: ids[1]});
 
-        if (!user1 || !user2) throw new BadRequestException("One or more of the users doesn't exist");
+        if (!user1 || !user2) throw new WsException("One or more of the users doesn't exist");
         if (user1.friends.includes(user2))
             return ;
         if(user1.blocked.includes(user2) || user2.blocked.includes(user1))
-            throw new BadRequestException("Users blocked each other");
+            throw new WsException("Users blocked each other");
         this.clients.get(ids[1]).emit("+" + ids[0]);
         this.clients.get(ids[0]).emit("+" + ids[1]);
         user1.friends.push(user2);
@@ -295,6 +275,9 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         const user = await this.userRepository.findOne({
             select: {
                 id: true,
+                channels: true,
+            },
+            relations: {
                 channels: true
             },
             where: {id: uid}
@@ -303,10 +286,11 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect 
         console.log(`chat user: ${user}`);
         if(!user)
             client.disconnect();
-        //client.emit(user.channels.toString());
+        client.emit("login", user);
         this.clients.set(user.id, client);
         this.sockets.set(client, user.id);
-        // client.join(user.channels.map(chan => chan.id.toString()));
+        console.log(user);
+        client.join(user.channels.map(chan => chan.id.toString()));
         this.logger.log('New client connected in chat gateway');
     }
